7th Semester/BDA (Big Data Analytics)/LAB/Amazon-S3/requirements.txt
Flask==2.2.3
pandas==1.5.3
boto3==1.26.7
Pillow==9.3.0
python-dotenv==0.21.0
Faker==15.3.2